requirements.txt
coloredlogs==15.0.1
PyYAML==5.4.1